refactor: move metadata readers
