+++ app/shell/py/pie/pie/build_index.py
@@ -10,9 +10,8 @@ import json
 import os
 from typing import Any, Dict, Optional
 
-import yaml
+from pie.metadata import read_from_markdown, read_from_yaml
 from pie.logging import logger, add_log_argument, setup_file_logger
-from pie.utils import read_yaml
 
 
 def get_url(filename: str) -> Optional[str]:
@@ -42,82 +41,6 @@ def get_url(filename: str) -> Optional[str]:
     raise Exception("Can't create a url.")
 
 
-def get_frontmatter(filename: str) -> Optional[Dict[str, Any]]:
-    """Extract YAML frontmatter from a Markdown file.
-
-    Args:
-        filename: Path to the Markdown file.
-
-    Returns:
-        A dict of the frontmatter if present; otherwise `None`.
-    """
-    with open(filename, encoding="utf-8") as file:
-        lines = file.readlines()
-
-    if not lines or lines[0].strip() != "---":
-        return None
-
-    yaml_lines = []
-    # Skip the first '---'
-    for line in lines[1:]:
-        if line.strip() == "---":
-            break
-        yaml_lines.append(line)
-
-    content = "".join(yaml_lines)
-    return yaml.safe_load(content)
-
-
-def process_markdown(filepath: str) -> Optional[Dict[str, Any]]:
-    """Load and prepare metadata from a Markdown file.
-
-    Extracts YAML frontmatter and, if applicable, computes the URL for the
-    Markdown document.
-
-    Args:
-        filepath: Path to the Markdown (`.md`) file.
-
-    Returns:
-        A metadata dict (including `url` if under `src/`) or `None` if
-        frontmatter is missing.
-    """
-    metadata = get_frontmatter(filepath)
-    if metadata is None:
-        logger.debug("No frontmatter found in Markdown file", filename=filepath)
-        return None
-    metadata["url"] = get_url(filepath)
-    return metadata
-
-
-def parse_yaml_metadata(filepath: str) -> Optional[Dict[str, Any]]:
-    """
-    Load and validate metadata from a YAML file.
-
-    Parses the YAML file, auto-generates missing fields (`citation`, `id`),
-    and computes the HTML URL if under `src/`.
-
-    Args:
-        filepath: Path to the YAML (`.yml`/`.yaml`) file.
-
-    Returns:
-        A metadata dict if parsing succeeds and content is a dict; otherwise
-        `None`.
-    """
-    try:
-        metadata = read_yaml(filepath)
-        if "url" not in metadata:
-            metadata["url"] = get_url(filepath)
-        if "citation" not in metadata:
-            # Intentionally use indexing so we get an exception here.
-            # The name field must exist.
-            metadata["citation"] = metadata["name"].lower()
-        if "id" not in metadata:
-            base, _ = os.path.splitext(filepath)
-            metadata["id"] = base.split(os.sep)[-1]
-        return metadata
-    except yaml.YAMLError:
-        logger.warning("Failed to parse YAML file", filename=filepath)
-        raise
 
 
 def validate_and_insert_metadata(
@@ -197,9 +120,9 @@ def build_index(
             metadata: Optional[Dict[str, Any]] = None
 
             if ext_lower == ".md":
-                metadata = process_markdown(filepath)
+                metadata = read_from_markdown(filepath)
             elif ext_lower in (".yml", ".yaml"):
-                metadata = parse_yaml_metadata(filepath)
+                metadata = read_from_yaml(filepath)
 
             if metadata:
                 validate_and_insert_metadata(metadata, filepath, index)
